Move EncryptionIfc definition to crypt_ifc module
